refactor(store_py): introduce PyClient and wrapper class for Python bindings
